requirements.txt
numpy
pandas
scikit-learn
matplotlib
hdbscan
torch
torchvision
seaborn
ipykernel
pillow
transformers
torchvision
